Structured project with go mod
