+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/skupperproject/skupper-example-tcp-echo
+
+go 1.15
